Reduce memory consumption of Docker images #76
The fix switches from an outdated Oracle Java 8 Docker image
to the current openjdk:8 image. The Java VM of this image
supports the parameters to make the Java VM Docker aware.
With this option it becomes possible to set the memory
limit to 256MB which reduces the memory consumption by
75% and leaves the application still operational.