Reduce memory consumption of Docker images #76 (#77)
* Reduce memory consumption of Docker images #76

The fix switches from an outdated Oracle Java 8 Docker image
to the current openjdk:8 image. The Java VM of this image
supports the parameters to make the Java VM Docker aware.
With this option it becomes possible to set the memory
limit to 256MB which reduces the memory consumption by
75% and leaves the application still operational.

* Added a NOTE about the Docker VM resource usage on Mac and Win

This closes #76 and hopefully improves the first time user-experience
of the Docker compose way of running the demo microservices.